git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/shelve-checkpoint@1817375 13f79535-47bb-0310-9956-ffa450edef68

+++ subversion/libsvn_client/shelve.c
@@ -547,7 +547,7 @@ svn_client_shelves_list(apr_hash_t **shelved_patch_infos,
     {
       const char *filename = apr_hash_this_key(hi);
       svn_io_dirent2_t *dirent = apr_hash_this_val(hi);
-      int len = strlen(filename);
+      size_t len = strlen(filename);
 
       if (len > 6 && strcmp(filename + len - 8, ".current") == 0)
         {
